refactor(publish-flow): extract restart/scale/rollback/eval mixins; facade under cap (Tasks 7-8)
Move restart, scale (+device-count), rollback+destroy, shared BaaS approve/record,
stage-status helpers, and eval publish/teardown into focused mixins (verbatim,
sharing self). Facade 2148 -> 996 lines; every publish_flow module now under the
1000-line cap, so the publish_flow_service.py allowlist entry is removed and the
architecture guard passes. 641 tests green. (Identifier renames still pending.)
