+++ src/backend/src/agentclaw/community/core/service_bot/services/publish_flow_service.py
@@ -6,13 +6,11 @@
 from __future__ import annotations
 
 import asyncio
-import time
 from collections.abc import Callable
-from typing import Any, Dict, TYPE_CHECKING
+from typing import TYPE_CHECKING
 
 from injector import inject
 
-from agentclaw.community.core.devices.models import DeviceBindingStatus
 from agentclaw.community.core.service_bot.repository.models import PublishStatus, BotPublishRecord
 from agentclaw.community.core.service_bot.schemas.publish_schemas import PublishFlowResult
 from agentclaw.community.core.service_bot.services.bot_build_service import BotBuildService
@@ -47,16 +45,30 @@ from agentclaw.community.core.service_bot.services.publish_flow.build_stage impo
 from agentclaw.community.core.service_bot.services.publish_flow.progress_sync_mixin import (
     ProgressSyncMixin,
 )
+from agentclaw.community.core.service_bot.services.publish_flow.restart_mixin import (
+    RestartMixin,
+)
+from agentclaw.community.core.service_bot.services.publish_flow.scale_mixin import (
+    ScaleMixin,
+)
+from agentclaw.community.core.service_bot.services.publish_flow.stage_status_mixin import (
+    StageStatusMixin,
+)
+from agentclaw.community.core.service_bot.services.publish_flow.rollback_ops_mixin import (
+    RollbackOpsMixin,
+)
+from agentclaw.community.core.service_bot.services.publish_flow.baas_publish_ops_mixin import (
+    BaasPublishOpsMixin,
+)
+from agentclaw.community.core.service_bot.services.publish_flow.eval_publish_mixin import (
+    EvalPublishMixin,
+)
 from agentclaw.community.core.service_bot.services.publish_flow.release_stage import (
     ONLINE_SPEC,
     VERIFY_SPEC,
     ReleaseStageRunner,
 )
 from agentclaw.community.log import get_logger
-from agentclaw.community.utils.env_utils import get_current_env
-
-DEVICE_COUNT_CONFIG_BUSINESS_CODE = "service_bot_device_count"
-DEVICE_COUNT_DEFAULT_PARAM_CODE = "default"
 
 
 if TYPE_CHECKING:
@@ -76,12 +88,16 @@ if TYPE_CHECKING:
 
 logger = get_logger()
 
-# ``PublishFlowServiceError`` now lives in ``publish_flow.errors`` and is imported
-# above; it stays importable from this module for backward compatibility
-# (services/__init__.py, router_publish.py, tests).
 
-
-class PublishFlowService(ProgressSyncMixin):
+class PublishFlowService(
+    ProgressSyncMixin,
+    RestartMixin,
+    ScaleMixin,
+    StageStatusMixin,
+    RollbackOpsMixin,
+    BaasPublishOpsMixin,
+    EvalPublishMixin,
+):
     """Bot 发布流程处理服务.
 
     职责：
@@ -948,932 +964,6 @@ class PublishFlowService(ProgressSyncMixin):
                 operator=operator,
             )
 
-    def restart_bot(
-        self,
-        publish_id: int,
-        operator: str = "system",
-    ) -> dict:
-        """重启 Bot（异步执行）。
-
-        根据发布单状态确定当前阶段，从 binding 信息获取 bot_uuid，调用 BaaS 层 upgrade 接口重新部署。
-        该方法使用 asyncio.create_task 异步执行，不等待结果。
-
-        流程：
-        1. 根据 publish_id 查询发布单记录
-        2. 根据发布单状态确定当前阶段（VERIFY/ONLINE）
-        3. 从 ext 获取对应阶段的 binding_id
-        4. 根据 binding_id 查询 device_binding 记录，获取 device_id（即 bot_uuid）
-        5. 调用 BotBuildService.upgrade_async() 重新部署 Bot
-
-        Args:
-            publish_id: 发布单 ID
-            operator: 操作者，默认为 "system"
-
-        Returns:
-            dict: 重启结果，包含:
-                - success: 是否成功提交重启请求
-                - message: 结果消息
-                - stage: 发布阶段（成功时返回）
-        """
-        logger.info(
-            f"[PublishFlowService.restart_bot] called: publish_id={publish_id}, operator={operator}"
-        )
-
-        # Step 1: 查询发布单记录
-        publish_record = self._publish_service.get_publish_by_id(publish_id)
-        if not publish_record:
-            logger.warning(
-                f"[PublishFlowService.restart_bot] Publish record not found: publish_id={publish_id}"
-            )
-            return {
-                "success": False,
-                "message": f"发布单不存在: publish_id={publish_id}",
-            }
-
-        # Step 2: 根据状态确定当前阶段
-        current_status = PublishStatus(publish_record.status)
-        stage = self._determine_restart_stage(current_status)
-        if not stage:
-            logger.warning(
-                f"[PublishFlowService.restart_bot] "
-                f"Cannot restart for status: {current_status}, publish_id={publish_id}"
-            )
-            return {
-                "success": False,
-                "message": f"当前状态 {current_status} 不支持重启操作",
-                "status": current_status,
-            }
-
-        # Step 3: 重置当前阶段的BaaS重启发布单
-        ext = self._get_latest_ext(publish_id)
-        if "restart" in ext and stage.value in ext.get("restart", {}):
-            try:
-                def _mutate(latest_ext: dict) -> None:
-                    restart_info = latest_ext.get("restart", {})
-                    if stage.value in restart_info:
-                        restart_info.pop(stage.value, None)
-                    if restart_info:
-                        latest_ext["restart"] = restart_info
-                    else:
-                        latest_ext.pop("restart", None)
-
-                ext = self._merge_and_update_ext(
-                    publish_id=publish_id,
-                    mutator=_mutate,
-                )
-            except Exception as e:
-                logger.warning(
-                    f"[PublishFlowService.restart_bot] "
-                    f"Failed to reset restart_publish_id: publish_id={publish_id}, error={e}"
-                )
-                ext = self._get_latest_ext(publish_id)
-
-        # Step 4: 从 ext 获取对应阶段的 binding_id
-        binding_info = ext.get("binding", {})
-        binding_id = binding_info.get(stage.value)
-
-        if not binding_id:
-            logger.warning(
-                f"[PublishFlowService.restart_bot] "
-                f"No binding_id found for stage={stage.value}, publish_id={publish_id}"
-            )
-            return {
-                "success": False,
-                "message": f"未找到 {stage.value} 阶段的绑定信息",
-                "stage": stage.value,
-            }
-
-        # Step 5: 根据 binding_id 查询 device_binding 记录
-        binding = self._publish_service.get_device_binding_by_id(binding_id)
-        if not binding:
-            logger.warning(
-                f"[PublishFlowService.restart_bot] Device binding not found: binding_id={binding_id}"
-            )
-            return {
-                "success": False,
-                "message": f"设备绑定记录不存在: binding_id={binding_id}",
-            }
-
-        bot_uuid = binding.device_id
-        if not bot_uuid:
-            logger.warning(
-                f"[PublishFlowService.restart_bot] No device_id in binding: binding_id={binding_id}"
-            )
-            return {
-                "success": False,
-                "message": f"设备绑定记录中没有 device_id: binding_id={binding_id}",
-            }
-
-        bot_service = self._bot_service
-        bot = bot_service.get_bot(bot_id=publish_record.source_bot_id, user_id=publish_record.owner_id)
-        if not bot:
-            logger.warning(
-                f"[PublishFlowService.restart_bot] Bot not found: bot_id={publish_record.source_bot_id}"
-            )
-            return {
-                "success": False,
-                "message": f"Bot不存在: {publish_record.source_bot_id}",
-            }
-
-        migration_path = ext.get("migration_path")
-        config_artifact = ext.get("config_artifact")
-        if not migration_path and not config_artifact:
-            logger.warning(
-                f"[PublishFlowService.restart_bot] No build artifact in ext: publish_id={publish_id}"
-            )
-            return {
-                "success": False,
-                "message": f"发布单缺少构建产物: publish_id={publish_id}",
-            }
-
-        # Step 6: 异步执行重启
-        asyncio.create_task(
-            self._restart_bot_async(
-                publish_id=publish_id,
-                publish_record=publish_record,
-                migration_path=migration_path,
-                bot_uuid=bot_uuid,
-                binding_id=binding_id,
-                bot=bot,
-                stage=stage,
-                operator=operator,
-            )
-        )
-
-        logger.info(
-            f"[PublishFlowService.restart_bot] Restart task created: "
-            f"publish_id={publish_id}, bot_uuid={bot_uuid}, stage={stage.value}, "
-            f"operator={operator}, owner_id={publish_record.owner_id}"
-        )
-
-        return {
-            "success": True,
-            "message": f"重启任务已提交，阶段: {stage.value}",
-            "stage": stage.value,
-            "bot_uuid": bot_uuid,
-        }
-
-    async def _restart_bot_async(
-        self,
-        publish_id: int,
-        publish_record: BotPublishRecord,
-        migration_path: str,
-        bot_uuid: str,
-        binding_id: int,
-        bot: Dict[str, Any],
-        stage: PublishStage,
-        operator: str,
-    ) -> None:
-        """异步执行 Bot 重启（通过 upgrade 接口重新部署）。
-
-        Args:
-            publish_id: 发布单 ID
-            publish_record: 发布单记录
-            migration_path: Bot 实例迁移后的目录路径
-            bot_uuid: Bot UUID
-            bot: Bot 信息字典
-            stage: 发布阶段
-            operator: 操作者
-        """
-        logger.info(
-            f"[PublishFlowService._restart_bot_async] Starting restart: "
-            f"publish_id={publish_id}, bot_uuid={bot_uuid}, stage={stage.value}, "
-            f"operator={operator}, owner_id={publish_record.owner_id}"
-        )
-
-        try:
-            # 生成 request_id（用于后续审批 BaaS 发布单）
-            request_id = self._build_service.generate_request_id(
-                bot=bot,
-                publish_stage=f"restart_{stage.value}",
-            )
-            version = f"{publish_record.version}"
-
-            # Compose the delivery artifact for THIS stage: stamp engine_ext.stage
-            # to the restarted stage and overlay that stage's stored channel
-            # engine_overrides (reproducing what was promoted, NOT a live re-fetch).
-            # Reading the per-stage slot fixes the single-config-slot hazard — a
-            # restart of a non-latest stage no longer delivers another stage's
-            # channels. No stored overrides (pre-feature record) or no
-            # config_artifact (ARCA) → no-ops, preserving prior behavior.
-            ext = publish_record.ext or {}
-            # `or {}` (not a get-default): the key may hold JSON null in a raw ext blob.
-            stored_overrides = (ext.get("engine_overrides_by_stage") or {}).get(stage.value)
-            config_artifact = self._artifact_for_stage(
-                ext.get("config_artifact"), stage, stored_overrides
-            )
-            restart_result = await self._build_service.upgrade_async(
-                bot_uuid=bot_uuid,
-                bot=bot,
-                user_id=publish_record.owner_id,
-                device_count=1,
-                migration_path=migration_path,
-                publish_stage=stage,
-                version=version,
-                config_artifact=config_artifact,
-            )
-
-            if restart_result.get("success") is False and restart_result.get("error_code") == "BOT_NOT_FOUND":
-                logger.warning(
-                    f"[PublishFlowService._restart_bot_async] "
-                    f"Restart target bot not found, fallback to first release: "
-                    f"publish_id={publish_id}, bot_uuid={bot_uuid}, stage={stage.value}"
-                )
-                restart_result = await self._build_service.release_async(
-                    bot=bot,
-                    user_id=publish_record.owner_id,
-                    migration_path=migration_path,
-                    device_count=1,
-                    publish_stage=stage,
-                    version=version,
-                    # teclaw: the fallback must carry the frozen artifact too,
-                    # else create_teclaw_bot would receive an empty config.
-                    config_artifact=config_artifact,
-                )
-
-            restart_publish_id = restart_result.get("publish_id")
-            if not restart_publish_id:
-                raise PublishFlowServiceError("BaaS 层重启未返回 publish_id")
-
-            # Refresh the reused binding's teclaw status read handle to the
-            # restart's publish workflow (no-op for non-teclaw; best-effort).
-            self._refresh_publish_handle(
-                binding_id, restart_publish_id
-            )
-
-            logger.info(
-                f"[PublishFlowService._restart_bot_async] "
-                f"Bot restart initiated: bot_uuid={bot_uuid}, stage={stage.value}, "
-                f"publish_id={publish_id}, restart_publish_id={restart_publish_id}"
-            )
-
-            # 将 restart_publish_id 存入 ext: {"restart": {"<stage>": restart_publish_id}}
-            try:
-                def _mutate(ext: dict) -> None:
-                    if "restart" not in ext:
-                        ext["restart"] = {}
-                    ext["restart"][stage.value] = restart_publish_id
-
-                self._merge_and_update_ext(
-                    publish_id=publish_id,
-                    mutator=_mutate,
-                )
-                logger.info(
-                    f"[PublishFlowService._restart_bot_async] "
-                    f"Restart publish_id saved to ext: publish_id={publish_id}, "
-                    f"stage={stage.value}, restart_publish_id={restart_publish_id}"
-                )
-            except Exception as save_error:
-                logger.warning(
-                    f"[PublishFlowService._restart_bot_async] "
-                    f"Failed to save restart_publish_id to ext: publish_id={publish_id}, "
-                    f"error={save_error}"
-                )
-
-            # 审批 BaaS 层重启发布单
-            self._approve_baas_publish(
-                baas_publish_id=restart_publish_id,
-                operator=operator,
-                stage=stage,
-                request_id=request_id,
-            )
-            logger.info(
-                f"[PublishFlowService._restart_bot_async] "
-                f"Bot restart approved: bot_uuid={bot_uuid}, stage={stage.value}, "
-                f"restart_publish_id={restart_publish_id}"
-            )
-
-        except Exception as e:
-            logger.error(
-                f"[PublishFlowService._restart_bot_async] "
-                f"Failed to restart bot: publish_id={publish_id}, bot_uuid={bot_uuid}, error={e}"
-            )
-
-    def scale_bot(
-        self,
-        publish_id: int,
-        operator: str = "system",
-    ) -> dict:
-        """对已发布的 service bot 发起扩缩容。
-
-        当前仅支持线上阶段（SUCCESS / ONLINE_PUB）的 service bot。
-        通过发布单线上 binding 获取 bot_uuid，然后调用 BaaS scale 接口。
-        """
-        logger.info(
-            f"[PublishFlowService.scale_bot] called: publish_id={publish_id}, "
-            f"operator={operator}"
-        )
-
-        publish_record = self._publish_service.get_publish_by_id(publish_id)
-        if not publish_record:
-            raise PublishNotFoundError(f"Publish order not found: {publish_id}")
-
-        owner_id = self._get_owner_id(publish_record)
-        bot = self._bot_service.get_bot(
-            bot_id=publish_record.source_bot_id,
-            user_id=owner_id,
-        )
-        if not bot:
-            raise PublishFlowServiceError(f"Bot不存在: {publish_record.source_bot_id}")
-
-        active_engine = (bot.get("active_engine") or "").strip().lower()
-        if not self._provider_behavior(bot).supports_scale:
-            return {
-                "success": True,
-                "message": "teclaw引擎的服务bot不支持扩容",
-                "publish_id": publish_id,
-                "engine": active_engine,
-                "supported": True,
-            }
-
-        current_status = PublishStatus(publish_record.status)
-        if current_status != PublishStatus.SUCCESS:
-            raise PublishStatusInvalidError(
-                f"当前状态 {current_status} 不支持扩容操作"
-            )
-
-        ext = self._get_latest_ext(publish_id)
-        online_binding_id = (ext.get("binding") or {}).get(PublishStage.ONLINE.value)
-        if not online_binding_id:
-            raise PublishFlowServiceError(
-                f"未找到 online 阶段的绑定信息: publish_id={publish_id}"
-            )
-
-        binding = self._publish_service.get_device_binding_by_id(online_binding_id)
-        if not binding or not binding.device_id:
-            raise PublishFlowServiceError(
-                f"设备绑定记录不存在或缺少 device_id: binding_id={online_binding_id}"
-            )
-
-        bot_uuid = binding.device_id
-        target_count = self._resolve_scale_target_count(publish_record)
-        request_id = f"scale_{bot_uuid}_{int(time.time() * 1000)}"
-
-        result = self._baas_service.scale_bot(
-            bot_uuid=bot_uuid,
-            owner_id=operator,
-            request_id=request_id,
-            target_count=target_count,
-            auto_approve_publish=True,
-        )
-
-        scale_publish_id = result.get("publish_id")
-        if scale_publish_id is not None:
-            def _mutate(latest_ext: dict) -> None:
-                latest_ext.setdefault("scale", {})["publish_id"] = scale_publish_id
-
-            self._merge_and_update_ext(
-                publish_id=publish_id,
-                mutator=_mutate,
-            )
-
-        logger.info(
-            f"[PublishFlowService.scale_bot] scale submitted: publish_id={publish_id}, "
-            f"bot_uuid={bot_uuid}, target_count={target_count}, result={result}"
-        )
-
-        return {
-            "success": True,
-            "message": "扩容任务已提交",
-            "publish_id": publish_id,
-            "bot_uuid": bot_uuid,
-            "target_count": target_count,
-            "baas_publish_id": result.get("publish_id"),
-            "data": result,
-        }
-
-    def _resolve_scale_target_count(self, publish_record: BotPublishRecord) -> int:
-        """解析服务 Bot 扩缩容目标设备数。"""
-        owner_id = self._get_owner_id(publish_record)
-        bot = self._bot_service.get_bot(
-            bot_id=publish_record.source_bot_id,
-            user_id=owner_id,
-        )
-        if not bot:
-            raise PublishFlowServiceError(f"Bot不存在: {publish_record.source_bot_id}")
-
-        ext_count = self._read_device_count_from_bot_ext(bot.get("ext"))
-        if ext_count is not None:
-            logger.info(
-                "[PublishFlowService._resolve_scale_target_count] use bot ext device_count=%s for bot_id=%s",
-                ext_count,
-                publish_record.source_bot_id,
-            )
-            return ext_count
-
-        default_count = self._get_default_scale_target_count()
-        if default_count is not None:
-            logger.info(
-                "[PublishFlowService._resolve_scale_target_count] use common_config default device_count=%s for bot_id=%s",
-                default_count,
-                publish_record.source_bot_id,
-            )
-            return default_count
-
-        raise PublishFlowServiceError(
-            f"未找到 service_bot_config.device_count，且未配置默认 device_count: bot_id={publish_record.source_bot_id}"
-        )
-
-    def _read_device_count_from_bot_ext(self, bot_ext: dict[str, Any] | None) -> int | None:
-        if not isinstance(bot_ext, dict):
-            return None
-        service_bot_config = bot_ext.get("service_bot_config")
-        if not isinstance(service_bot_config, dict):
-            return None
-        return self._normalize_device_count(
-            service_bot_config.get("device_count"),
-            source="bot.ext.service_bot_config.device_count",
-        )
-
-    def _get_default_scale_target_count(self) -> int | None:
-        value = self._common_config_service.get_value(
-            business_code=DEVICE_COUNT_CONFIG_BUSINESS_CODE,
-            param_code=DEVICE_COUNT_DEFAULT_PARAM_CODE,
-            env=get_current_env(),
-            default=None,
-        )
-        logger.info(
-            "[PublishFlowService._get_default_scale_target_count] common_config value=%s",
-            value,
-        )
-        return self._normalize_device_count(
-            value,
-            source="common_config.service_bot_device_count.default",
-        )
-
-    def _normalize_device_count(self, value: Any, *, source: str) -> int | None:
-        if value is None or value == "":
-            return None
-        try:
-            count = int(value)
-        except (TypeError, ValueError):
-            logger.warning(
-                "[PublishFlowService._normalize_device_count] invalid device_count from %s: %r",
-                source,
-                value,
-            )
-            return None
-        if count < 1:
-            logger.warning(
-                "[PublishFlowService._normalize_device_count] non-positive device_count from %s: %r",
-                source,
-                value,
-            )
-            return None
-        return count
-
-    def _determine_restart_stage(self, current_status: PublishStatus) -> PublishStage | None:
-        """根据当前状态确定重启阶段。
-
-        Args:
-            current_status: 当前发布单状态
-
-        Returns:
-            发布阶段（VERIFY/ONLINE），如果不支持重启则返回 None
-        """
-        # VALIDATING / VALIDATE_PUB: 验证环境，重启验证环境的 bot
-        if current_status in (PublishStatus.VALIDATING, PublishStatus.VALIDATE_PUB):
-            return PublishStage.VERIFY
-        # SUCCESS / ONLINE_PUB: 线上环境，重启线上的 bot
-        elif current_status in (PublishStatus.SUCCESS, PublishStatus.ONLINE_PUB):
-            return PublishStage.ONLINE
-        return None
-
-    def _determine_sync_stage(self, current_status: PublishStatus) -> PublishStage | None:
-        """根据当前状态确定同步阶段。
-
-        Args:
-            current_status: 当前发布单状态
-
-        Returns:
-            发布阶段（VERIFY/ONLINE），如果不支持同步则返回 None
-        """
-        if current_status == PublishStatus.VALIDATE_PUB:
-            return PublishStage.VERIFY
-        elif current_status == PublishStatus.ONLINE_PUB:
-            return PublishStage.ONLINE
-        return None
-
-    def _upgrade_last_publish(
-        self,
-        publish_record: BotPublishRecord,
-        stage: PublishStage,
-        target_status: PublishStatus,
-    ) -> None:
-        """更新上一个发布单状态为 UPGRADED（仅当 online 阶段成功时）。
-
-        Args:
-            publish_record: 当前发布单记录
-            stage: 发布阶段（VERIFY/ONLINE）
-            target_status: 目标状态
-        """
-        # 只有 online 阶段成功时才需要更新上一个发布单
-        if stage != PublishStage.ONLINE or target_status != PublishStatus.SUCCESS:
-            return
-
-        last_pub_id = publish_record.last_pub_id
-        if not last_pub_id or last_pub_id <= 0:
-            return
-
-        # 查询上一个发布单记录
-        last_publish = self._publish_service.get_publish_by_id(last_pub_id)
-        if not last_publish:
-            logger.warning(
-                f"[PublishFlowService._upgrade_last_publish] "
-                f"Last publish record not found: last_pub_id={last_pub_id}"
-            )
-            return
-
-        # 清除 rollback_restored_from 标记（如果存在）
-        last_ext = last_publish.ext or {}
-        if last_ext.pop("rollback_restored_from", None):
-            logger.info(
-                f"[PublishFlowService._upgrade_last_publish] "
-                f"Clearing rollback_restored_from for publish {last_pub_id}"
-            )
-
-        # 更新上一个发布单状态为 UPGRADED，同时更新 ext
-        try:
-            self._publish_service.update_publish_status_with_ext(
-                publish_id=last_pub_id,
-                target_status=PublishStatus.UPGRADED,
-                ext=last_ext,
-                source_status=PublishStatus.SUCCESS,
-            )
-            logger.info(
-                f"[PublishFlowService._upgrade_last_publish] "
-                f"Last publish status updated to UPGRADED: last_pub_id={last_pub_id}"
-            )
-        except Exception as e:
-            logger.warning(
-                f"[PublishFlowService._upgrade_last_publish] "
-                f"Failed to update last publish status: last_pub_id={last_pub_id}, error={e}"
-            )
-
-    async def execute_rollback(
-        self,
-        current_publish_id: int,
-        target_publish_id: int,
-        operator: str,
-    ) -> "PublishFlowResult":
-        """执行回滚部署。
-
-        使用目标版本的配置（migration_path/config_artifact/binding）重新部署到线上。
-        回滚部署在目标版本上进行，前端应同步 target_publish_id 的部署进度。
-
-        Args:
-            current_publish_id: 当前版本 ID（已变为 DRAFT，仅用于获取 owner_id 和 bot_id）
-            target_publish_id: 目标版本 ID（回滚目标，已变为 SUCCESS，部署在此版本上）
-            operator: 操作者
-
-        Returns:
-            PublishFlowResult: 部署结果，publish_id 为 target_publish_id
-        """
-        from agentclaw.community.core.service_bot.schemas.publish_schemas import PublishFlowResult
-
-        logger.info(
-            f"[PublishFlowService.execute_rollback] called: "
-            f"current_publish_id={current_publish_id}, target_publish_id={target_publish_id}, "
-            f"operator={operator}"
-        )
-
-        # 1. 获取目标版本记录
-        target_record = self._publish_service.get_publish_by_id(target_publish_id)
-        if not target_record:
-            raise PublishNotFoundError(f"Target publish record not found: {target_publish_id}")
-
-        current_record = self._publish_service.get_publish_by_id(current_publish_id)
-        if not current_record:
-            raise PublishNotFoundError(f"Current publish record not found: {current_publish_id}")
-
-        # 2. 获取目标版本的构建产物
-        target_ext = self._get_latest_ext(target_publish_id)
-        migration_path = target_ext.get("migration_path")
-        config_artifact = target_ext.get("config_artifact")
-
-        if not migration_path and not config_artifact:
-            raise PublishFlowServiceError(
-                f"目标版本缺少构建产物: target_publish_id={target_publish_id}"
-            )
-
-        # 3. 从目标版本获取线上 binding
-        online_binding_id = target_ext.get("binding", {}).get(PublishStage.ONLINE.value)
-        if not online_binding_id:
-            raise PublishFlowServiceError(
-                f"目标版本缺少线上 binding: target_publish_id={target_publish_id}"
-            )
-
-        binding = self._publish_service.get_device_binding_by_id(online_binding_id)
-        if not binding or not binding.device_id:
-            raise PublishFlowServiceError(
-                f"设备绑定记录不存在或缺少 device_id: binding_id={online_binding_id}"
-            )
-
-        bot_uuid = binding.device_id
-
-        # 4. 获取 Bot 信息（从 current_record 获取，因为它是原始发布单）
-        owner_id = current_record.owner_id
-        bot = self._bot_service.get_bot(bot_id=current_record.source_bot_id, user_id=owner_id)
-        if not bot:
-            raise PublishFlowServiceError(f"Bot不存在: {current_record.source_bot_id}")
-
-        # 5. 调用 BaaS upgrade 接口重新部署
-        version = f"{target_record.version}"
-        upgrade_result = await self._build_service.upgrade_async(
-            bot_uuid=bot_uuid,
-            bot=bot,
-            user_id=owner_id,
-            device_count=1,
-            migration_path=migration_path,
-            publish_stage=PublishStage.ONLINE,
-            version=version,
-            config_artifact=config_artifact,
-        )
-
-        baas_publish_id = upgrade_result.get("publish_id")
-        if not baas_publish_id:
-            raise PublishFlowServiceError("BaaS 层升级未返回 publish_id")
-
-        # 6. 更新目标版本记录新的 BaaS 发布单 ID（回滚部署在目标版本上进行）
-        if "publish" not in target_ext:
-            target_ext["publish"] = {}
-        target_ext["publish"][PublishStage.ONLINE.value] = baas_publish_id
-
-        self._update_publish_status(
-            publish_id=target_publish_id,
-            target_status=PublishStatus.ONLINE_PUB,
-            source_status=PublishStatus.SUCCESS,
-            ext=target_ext,
-        )
-
-        # 7. 审批 BaaS 发布单
-        request_id = self._build_service.generate_request_id(
-            bot=bot,
-            publish_stage="rollback",
-        )
-        self._approve_baas_publish(
-            baas_publish_id=baas_publish_id,
-            operator=operator,
-            stage=PublishStage.ONLINE,
-            request_id=request_id,
-        )
-
-        logger.info(
-            f"[PublishFlowService.execute_rollback] Rollback deployment initiated: "
-            f"current_publish_id={current_publish_id}, target_publish_id={target_publish_id}, "
-            f"bot_uuid={bot_uuid}, baas_publish_id={baas_publish_id}"
-        )
-
-        return PublishFlowResult(
-            publish_id=target_publish_id,
-            status=PublishStatus.ONLINE_PUB,
-            message="回滚发布已提交",
-            action="rollback",
-            target_publish_id=target_publish_id,
-            bot_uuid=bot_uuid,
-            baas_publish_id=str(baas_publish_id),
-            device_binding_id=online_binding_id,
-        )
-
-    def _destroy_bot_by_stage(
-        self,
-        publish_record: BotPublishRecord,
-        stage: PublishStage,
-    ) -> None:
-        """销毁指定阶段的 bot 实例。
-
-        Args:
-            publish_record: 发布单记录
-            stage: 发布阶段（VERIFY/ONLINE）
-        """
-        ext = publish_record.ext or {}
-        binding_info = ext.get("binding", {})
-        binding_id = binding_info.get(stage.value)
-
-        if not binding_id:
-            logger.warning(
-                f"[PublishFlowService._destroy_bot_by_stage] "
-                f"No binding_id found for stage={stage.value}, skipping destroy"
-            )
-            return
-
-        try:
-            # 查询 device_binding 获取 bot_uuid
-            binding = self._publish_service.get_device_binding_by_id(binding_id)
-            if not binding:
-                logger.warning(
-                    f"[PublishFlowService._destroy_bot_by_stage] "
-                    f"Device binding not found: binding_id={binding_id}"
-                )
-                return
-
-            bot_uuid = binding.device_id
-            if not bot_uuid:
-                logger.warning(
-                    f"[PublishFlowService._destroy_bot_by_stage] "
-                    f"No device_id in binding: binding_id={binding_id}"
-                )
-                return
-
-            logger.info(
-                f"[PublishFlowService._destroy_bot_by_stage] "
-                f"Destroying bot: bot_uuid={bot_uuid}, stage={stage.value}"
-            )
-
-            # 生成 request_id（销毁场景使用特殊标识）
-            request_id = self._build_service.generate_request_id(
-                bot={"entity_id": binding.entity_id, "entity_type": binding.entity_type, "bot_id": publish_record.source_bot_id},
-                publish_stage=f"destroy_{stage.value}",
-            )
-
-            # 调用 BaaS 销毁 bot
-            destroy_result = self._baas_service.stop_bot(
-                bot_uuid=bot_uuid,
-                operator="system",
-                request_id=request_id,
-            )
-
-            destroy_publish_id = destroy_result.get("publish_id")
-            logger.info(
-                f"[PublishFlowService._destroy_bot_by_stage] "
-                f"Bot destroy initiated: bot_uuid={bot_uuid}, stage={stage.value}, destroy_publish_id={destroy_publish_id}"
-            )
-
-            # 审批销毁流程单
-            if destroy_publish_id:
-                self._approve_baas_publish(
-                    baas_publish_id=destroy_publish_id,
-                    operator="system",
-                    stage=stage,
-                    request_id=request_id,
-                )
-                logger.info(
-                    f"[PublishFlowService._destroy_bot_by_stage] "
-                    f"Bot destroy approved: bot_uuid={bot_uuid}, stage={stage.value}, destroy_publish_id={destroy_publish_id}"
-                )
-
-            # 更新 device_binding 状态为 RELEASED
-            self._publish_service.update_device_binding_with_props(
-                binding_id=binding_id,
-                status=DeviceBindingStatus.RELEASED,
-                device_props={"destroy_publish_id": destroy_publish_id} if destroy_publish_id else {},
-            )
-            logger.info(
-                f"[PublishFlowService._destroy_bot_by_stage] "
-                f"Device binding status updated to RELEASED: binding_id={binding_id}"
-            )
-
-        except Exception as e:
-            logger.warning(
-                f"[PublishFlowService._destroy_bot_by_stage] "
-                f"Failed to destroy bot: binding_id={binding_id}, stage={stage.value}, error={e}"
-            )
-
-
-    def _approve_baas_publish(
-        self,
-        baas_publish_id: int | str,
-        operator: str,
-        stage: PublishStage,
-        request_id: str,
-    ) -> bool:
-        """审批 BaaS 层发布单。
-
-        Args:
-            baas_publish_id: BaaS 层发布工作流 ID
-            operator: 审批操作者用户 ID
-            stage: 发布阶段，用于日志区分
-            request_id: 请求 ID，用于幂等性控制
-        """
-        if not baas_publish_id:
-            logger.warning(
-                f"[PublishFlowService._approve_baas_publish] "
-                f"No baas_publish_id provided, skipping approve for stage={stage.value}"
-            )
-            return False
-
-        logger.info(
-            f"[PublishFlowService._approve_baas_publish] "
-            f"Approving BaaS publish: baas_publish_id={baas_publish_id}, stage={stage.value}"
-        )
-
-        try:
-            self._baas_service.approve_publish(
-                publish_id=int(baas_publish_id),
-                operator=operator,
-                request_id=request_id,
-                comment=f"自动审批 - {stage.value}环境发布",
-            )
-            logger.info(
-                f"[PublishFlowService._approve_baas_publish] "
-                f"BaaS publish approved: baas_publish_id={baas_publish_id}, stage={stage.value}"
-            )
-            return True
-        except Exception as e:
-            logger.warning(
-                f"[PublishFlowService._approve_baas_publish] "
-                f"Failed to approve BaaS publish: baas_publish_id={baas_publish_id}, "
-                f"stage={stage.value}, error={e}, continuing..."
-            )
-            return False
-
-    def _record_release_result(
-        self,
-        publish_id: int,
-        bot: dict,
-        bot_uuid: str,
-        baas_publish_id: int | str,
-        operator: str,
-        ext: dict,
-        stage: PublishStage,
-        source_status: PublishStatus,
-        target_status: PublishStatus,
-        engine_overrides: dict | None = None,
-    ) -> tuple[int, dict]:
-        """记录发布结果：创建 device binding 并更新扩展字段。
-
-        Args:
-            publish_id: 发布单 ID
-            bot: Bot 信息字典
-            bot_uuid: BaaS 层返回的 bot_uuid
-            baas_publish_id: BaaS 层发布单 ID
-            operator: 操作者
-            ext: 扩展字段
-            stage: 发布阶段（VERIFY/ONLINE）
-            source_status: 源状态
-            target_status: 目标状态
-
-        Returns:
-            tuple[int, dict]: (binding_id, 更新后的 ext)
-        """
-        # 创建 device binding 记录。device_provider 由 baas 决定（查源 bot 的容器
-        # provider_type，teclaw→teclaw 否则 baas），替代原先写死的 "baas"；与 build
-        # 阶段的生产者选择共用同一解析规则（同一 resolve_container_provider）。
-        binding_id = self._publish_service.create_device_binding(
-            entity_id=bot.get("entity_id", ""),
-            entity_type=bot.get("entity_type", "staff"),
-            device_id=bot_uuid,
-            device_provider=self._baas_service.resolve_container_provider(bot),
-            # Stash the baas publish_id as the teclaw status read handle (no-op
-            # for non-teclaw bindings — their status reads ignore this key).
-            device_props={"publish_id": baas_publish_id},
-            applied_by=operator,
-            apply_reason="",
-            status=DeviceBindingStatus.PENDING,
-        )
-
-        ext = self._get_latest_ext(publish_id)
-
-        # 记录 binding_id 到扩展字段
-        if "binding" not in ext:
-            ext["binding"] = {}
-        ext["binding"][stage.value] = binding_id
-
-        # 记录 baas 层发布单 ID 到扩展字段
-        if "publish" not in ext:
-            ext["publish"] = {}
-        ext["publish"][stage.value] = baas_publish_id
-
-        # 把本阶段（canary/release）持久化进存储的 config_artifact 快照。本方法在内部
-        # 重新读取了 ext（覆盖调用方的修改），所以首次发布路径的 stage 持久化必须落在
-        # 这里；ARCA 无 config_artifact 时 no-op。
-        self._restamp_ext_artifact(ext, stage)
-
-        # Persist this stage's engine_overrides (DingTalk channels) next to the
-        # binding/publish refs. Same reason as the restamp above: this method
-        # re-reads ext, so the first-release store must land here. No-op for ARCA
-        # (engine_overrides is None).
-        self._store_stage_overrides(ext, stage, engine_overrides)
-
-        # 更新状态
-        self._update_publish_status(
-            publish_id=publish_id,
-            target_status=target_status,
-            source_status=source_status,
-            ext=ext,
-        )
-
-        # 审批 BaaS 层发布单
-        request_id = self._build_service.generate_request_id(
-            bot=bot,
-            publish_stage=stage.value,
-        )
-
-        self._approve_baas_publish(
-            baas_publish_id=baas_publish_id,
-            operator=operator,
-            stage=stage,
-            request_id=request_id,
-        )
-
-        logger.info(
-            f"[PublishFlowService._record_release_result] "
-            f"Release recorded: publish_id={publish_id}, stage={stage.value}, "
-            f"binding_id={binding_id}, baas_publish_id={baas_publish_id}"
-        )
-
-        return binding_id, ext
-
     @staticmethod
     def _restamp_ext_artifact(ext: dict, stage: PublishStage) -> None:
         PublishExtState.stamp_stage_on_stored_artifact(ext, stage)
@@ -1904,248 +994,3 @@ class PublishFlowService(ProgressSyncMixin):
     ) -> None:
         self._ext_state.update_status(publish_id, target_status, source_status, ext)
 
-    async def general_publish(
-        self,
-        publish_id: int,
-        operator: str,
-        publish_stage: PublishStage = PublishStage.EVAL,
-        biz_id: str = "",
-    ) -> dict:
-        """发布评估环境。
-
-        评估环境是主发布流程之外的旁支能力：
-        - 不推进主发布单状态机
-        - 不写入 publish.ext / binding
-        - 仅复用发布单上的构建产物与 bot 基础信息
-        """
-        logger.info(
-            f"[PublishFlowService.general_publish] Start release: "
-            f"publish_id={publish_id}, operator={operator}"
-        )
-
-        publish_record = self._publish_service.get_publish_by_id(publish_id)
-        if not publish_record:
-            raise PublishNotFoundError(f"Publish order not found: {publish_id}")
-
-        owner_id = self._get_owner_id(publish_record)
-        migration_path = (publish_record.ext or {}).get("migration_path", "")
-        config_artifact = (publish_record.ext or {}).get("config_artifact")
-        if not migration_path and not config_artifact:
-            raise PublishFlowServiceError("构建产物路径不存在，请先执行构建")
-
-        bot = self._bot_service.get_bot(
-            bot_id=publish_record.source_bot_id,
-            user_id=owner_id,
-        )
-        if not bot:
-            raise PublishFlowServiceError(f"Bot不存在: {publish_record.source_bot_id}")
-
-        eval_overrides = self._stage_overrides(publish_record, publish_stage)
-        eval_artifact = self._artifact_for_stage(
-            config_artifact,
-            publish_stage,
-            eval_overrides,
-        )
-
-        ext_info = {}
-        if biz_id:
-            ext_info["biz_id"] = biz_id
-
-        # 发布评估环境
-        release_result = await self._build_service.release_async(
-            bot=bot,
-            user_id=owner_id,
-            migration_path=migration_path,
-            device_count=1,
-            publish_stage=publish_stage,
-            version=str(publish_record.version or 1),
-            config_artifact=eval_artifact,
-            ext_info=ext_info,
-        )
-
-        bot_uuid = release_result.get("bot_uuid")
-        baas_publish_id = release_result.get("publish_id")
-        if not bot_uuid:
-            raise PublishFlowServiceError("评估环境发布失败: BaaS 未返回 bot_uuid")
-
-        request_id = self._build_service.generate_request_id(
-            bot=bot,
-            publish_stage=publish_stage.value,
-        )
-        self._approve_baas_publish(
-            baas_publish_id=baas_publish_id,
-            operator=operator,
-            stage=publish_stage,
-            request_id=request_id,
-        )
-
-        result = {
-            "success": True,
-            "publish_id": publish_id,
-            "stage": publish_stage.value,
-            "bot_uuid": bot_uuid,
-            "baas_publish_id": baas_publish_id,
-            "baas_bot_status": release_result.get("status"),
-        }
-        logger.info(
-            f"[PublishFlowService.general_publish] Release success: {result}"
-        )
-        return result
-
-    def general_teardown(
-        self,
-        bot_uuid: str,
-        *,
-        operator: str = "system",
-        request_bot: dict | None = None,
-    ) -> dict:
-        """销毁评估环境。
-
-        仅依赖评估环境自身 bot_uuid；调用方后续可改为从独立评估任务表读取并传入。
-        不触碰主发布单 ext/binding。
-        """
-        if not bot_uuid:
-            raise PublishFlowServiceError("bot_uuid 不能为空")
-
-        request_bot = request_bot or {"bot_id": bot_uuid}
-        request_id = self._build_service.generate_request_id(
-            bot=request_bot,
-            publish_stage="destroy_eval",
-        )
-        destroy_result = self._baas_service.destroy_bot(
-            bot_uuid=bot_uuid,
-            operator=operator,
-            request_id=request_id,
-        )
-        destroy_publish_id = destroy_result.get("publish_id")
-        self._approve_baas_publish(
-            baas_publish_id=destroy_publish_id,
-            operator=operator,
-            stage=PublishStage.EVAL,
-            request_id=request_id,
-        )
-        result = {
-            "success": True,
-            "bot_uuid": bot_uuid,
-            "baas_publish_id": destroy_publish_id,
-            "message": "评估环境销毁已提交",
-        }
-        logger.info(
-            f"[PublishFlowService.general_teardown] Destroy success: {result}"
-        )
-        return result
-
-    def get_publish_bot_status(
-        self,
-        publish_id: int,
-        stage: PublishStage,
-    ) -> Dict[str, Any]:
-        """查询指定发布单阶段对应的 BaaS bot 状态 / 详情。"""
-        logger.info(
-            f"[PublishFlowService.get_publish_bot_status] Start query: publish_id={publish_id}, stage={stage.value}"
-        )
-
-        publish_record = self._publish_service.get_publish_by_id(publish_id)
-        if not publish_record:
-            raise PublishNotFoundError(f"Publish order not found: {publish_id}")
-
-        ext = publish_record.ext or {}
-        binding_info = ext.get("binding", {})
-        binding_id = binding_info.get(stage.value)
-        if not binding_id:
-            raise PublishFlowServiceError(f"未找到 {stage.value} 阶段的绑定信息")
-
-        binding = self._publish_service.get_device_binding_by_id(binding_id)
-        if not binding:
-            raise PublishFlowServiceError(f"未找到绑定记录: binding_id={binding_id}")
-
-        bot_uuid = getattr(binding, "device_id", "")
-        if not bot_uuid:
-            raise PublishFlowServiceError(f"绑定记录缺少 bot_uuid: binding_id={binding_id}")
-
-        baas_bot = self._baas_service.get_bot(bot_uuid=bot_uuid)
-        result = {
-            "publish_id": publish_id,
-            "stage": stage.value,
-            "binding_id": binding_id,
-            "bot_uuid": bot_uuid,
-            "baas_bot_status": baas_bot.get("status"),
-        }
-        logger.info(
-            f"[PublishFlowService.get_publish_bot_status] Query success: {result}"
-        )
-        return result
-
-    def destroy_publish_history(
-        self,
-        publish_id: int,
-        stage: PublishStage,
-    ) -> dict:
-        """销毁发布历史。
-
-        注意：调用此方法前需先将 bot 推进到相应状态：
-        - 验证阶段 (VERIFY): 需先回退到草稿状态 (DRAFT)
-        - 线上阶段 (ONLINE): 需先推进到下线状态 (RELEASED)
-
-        销毁流程：
-        1. 通过 publish_id 查询 BotPublishRecord 记录
-        2. 调用 _destroy_bot_by_stage 方法销毁指定阶段的 BaaS 层 bot
-
-        Args:
-            publish_id: AgentClaw 层发布单 ID
-            stage: 发布阶段（VERIFY/ONLINE）
-
-        Returns:
-            dict: 销毁结果，包含:
-                - success: 是否成功
-                - bot_destroyed: 是否销毁了 bot
-                - message: 结果消息
-
-        Raises:
-            PublishNotFoundError: 发布单不存在
-            PublishFlowServiceError: 销毁失败
-        """
-        logger.info(f"[PublishFlowService.destroy_publish_history] Starting destroy: publish_id={publish_id}, stage={stage.value}")
-
-        # Step 1: 查询发布单
-        publish_record = self._publish_service.get_publish_by_id(publish_id)
-        if not publish_record:
-            raise PublishNotFoundError(f"Publish order not found: {publish_id}")
-
-        current_status = PublishStatus(publish_record.status)
-
-        # 检查状态：只有 DRAFT 和 RELEASED 状态才能销毁
-        allowed_statuses = [PublishStatus.DRAFT, PublishStatus.RELEASED]
-        if current_status not in allowed_statuses:
-            raise PublishFlowServiceError(
-                f"发布单状态不允许销毁: 当前状态={current_status}, "
-                f"仅允许状态: {[s.value for s in allowed_statuses]}"
-            )
-
-        result = {
-            "success": True,
-            "bot_destroyed": False,
-            "message": "",
-        }
-
-        # Step 2: 销毁 BaaS 层的 bot
-        try:
-            self._destroy_bot_by_stage(publish_record, stage)
-            result["bot_destroyed"] = True
-            logger.info(
-                f"[PublishFlowService.destroy_publish_history] "
-                f"Bot destroyed: publish_id={publish_id}, stage={stage.value}"
-            )
-
-        except Exception as e:
-            logger.warning(
-                f"[PublishFlowService.destroy_publish_history]"
-                f"Failed to destroy BaaS bots: publish_id={publish_id}, stage={stage.value}, error={e}"
-            )
-            # 销毁 bot 失败不阻塞整体流程
-
-        result["message"] = f"发布历史销毁完成: publish_id={publish_id}, stage={stage.value}"
-
-        logger.info(f"[PublishFlowService.destroy_publish_history] Destroy completed: {result}")
-
-        return result
